requirements.txt
# Python 核心依賴
python-binance==1.0.32
ccxt==4.5.17
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
pyarrow>=14.0.0
tqdm>=4.66.0

# 技術指標
TA-Lib>=0.4.28

# 機器學習
xgboost>=2.0.0
lightgbm>=4.1.0
scikit-learn>=1.3.0
optuna>=3.5.0
shap>=0.43.0

# 時間序列
prophet>=1.1.5
statsmodels>=0.14.0

# 資料庫
# psycopg2-binary==2.9.9
# influxdb-client==1.39.0
# redis==5.0.1
# sqlalchemy==2.0.23

# Web 框架
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0

# 資料處理
pydantic==2.5.3
pydantic-settings==2.1.0

# 任務佇列
celery==5.3.4
flower==2.0.1

# 監控
prometheus-client==0.19.0

# 視覺化
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# 日誌
loguru==0.7.2

# 測試
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# 代碼品質
black==23.12.1
flake8==7.0.0
mypy==1.7.1
pre-commit==3.6.0

# 工具
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
